milestone4/requirements.txt
mlflow>=2.8.0,<3.0.0
fastapi>=0.100.0,<0.105.0
uvicorn>=0.23.0,<0.25.0
python-dotenv>=1.0.0,<2.0.0
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<1.27.0
scikit-learn>=1.3.0,<1.4.0
prometheus-client>=0.17.0,<0.20.0
evidently>=0.3.0,<0.5.0
python-multipart>=0.0.6,<0.1.0
pydantic>=2.0.0,<3.0.0
joblib>=1.3.0,<1.4.0
requests>=2.31.0,<3.0.0 